Reorganize to src folder
