--- index.js
@@ -1,5 +1,0 @@
-exports.list = require('./src/list');
-exports.load = require('./src/load');
-exports.stats = require('./src/stats');
-exports.validate = require('./src/validate');
-exports.dependencies = require('./src/dependencies');
