+++ package.json
@@ -2,7 +2,7 @@
   "name": "fbp-manifest",
   "version": "0.2.7",
   "description": "Flow-Based Programming Manifest tools",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "nyc grunt test"
   },
@@ -45,9 +45,9 @@
   },
   "nyc": {
     "include": [
-      "index.js",
       "src/*.js",
-      "src/**/*.js"
+      "src/**/*.js",
+      "src/**/**/*.js"
     ]
   }
 }
