+++ src/index.js
@@ -1,0 +1,5 @@
+exports.list = require('./lib/list');
+exports.load = require('./lib/load');
+exports.stats = require('./lib/stats');
+exports.validate = require('./lib/validate');
+exports.dependencies = require('./lib/dependencies');
