+++ src/lib/dependencies.js
