+++ src/lib/stats.js
